Unhandle NO and BAD unsolicited responses.
By not handling these we ensure they will turn into an Err.
These responses should only be returned as a solicited response, so handing
them to a IDLE waiter as an unsolicited responses makes the user implement
correct protocol handling of them, when they should always be treated as an
error.